requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
streamlit>=1.31.0
joblib>=1.3.0
scipy>=1.10.0

# Operations research
pulp>=2.7.0

# Graph analysis
networkx>=3.0

# REST API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Azure — Data Lake Storage Gen2 (Hierarchical Namespace)
azure-storage-file-datalake>=12.0.0
azure-storage-blob>=12.0.0

# Azure — Authentication (DefaultAzureCredential — no hardcoded keys)
azure-identity>=1.15.0

# Azure — Key Vault (secrets management)
azure-keyvault-secrets>=4.7.0

# MLflow — experiment tracking (Azure ML uses MLflow natively)
mlflow>=2.10.0
python-dotenv>=1.0.0
